+++ lib/xirr/config.rb
@@ -6,7 +6,8 @@ module Xirr
       eps: '1.0e-12',
       days_in_year: 365,
       iteration_limit: 100,
-      precision: 6
+      precision: 6,
+      default_method: :bisection
   }
 
   # Iterates trhough default values and sets in config